Fix: Added error toast message for login if user is not verified yet

+++ feature/auth/src/commonMain/kotlin/com/creativehazio/auth/presentation/login/LoginScreen.kt
@@ -67,14 +67,14 @@ fun LoginScreenRoot(
                 is LoginEffect.ShowError -> {
                     snackbarHostState.showSnackbar(
                         message = it.error.message().asStringSuspend(),
-                        duration = SnackbarDuration.Long
+                        duration = SnackbarDuration.Short
                     )
                 }
 
                 is LoginEffect.ShowSuccess -> {
                     snackbarHostState.showSnackbar(
                         message = it.message,
-                        duration = SnackbarDuration.Long
+                        duration = SnackbarDuration.Short
                     )
                 }
